// inmo/src/app/interfaces/inversion/inversionInterface.ts
export interface InversionInterface {
    idinversion: number;
    usuario_idusuario: number;
    proyecto_inmobiliario_idproyecto_inmobiliario: number;
    cantidad: number;
    fecha_inversion: Date;
    estado_inversion: string;
    contrato_participacion_cliente: string;
    contrato_participacion_ambos: string;
    recibo_firmado: string;
    fecha_recibo: Date;
}